chore: have resolver automatically inject stdlib
