+++ crates/nargo/src/cli/compile_cmd.rs
@@ -11,7 +11,7 @@ use crate::{constants::TARGET_DIR, errors::CliError, resolver::Resolver};
 
 use super::fs::program::save_program_to_file;
 use super::preprocess_cmd::preprocess_with_path;
-use super::{add_std_lib, NargoConfig};
+use super::NargoConfig;
 
 /// Compile the program and its secret execution trace into ACIR format
 #[derive(Debug, Clone, Args)]
@@ -69,9 +69,7 @@ pub(crate) fn run(args: CompileCommand, config: NargoConfig) -> Result<(), CliEr
 
 fn setup_driver(program_dir: &Path) -> Result<Driver, CliError> {
     let backend = crate::backends::ConcreteBackend;
-    let mut driver = Resolver::resolve_root_config(program_dir, backend.np_language())?;
-    add_std_lib(&mut driver);
-    Ok(driver)
+    Resolver::resolve_root_config(program_dir, backend.np_language())
 }
 
 fn check_crate(program_dir: &Path, options: &CompileOptions) -> Result<Driver, CliError> {
